Updated Home Carousel.
