+++ src/templates/main/index.html
@@ -35,7 +35,7 @@
                 <!--First slide-->
                 {% for image in carousel %}
                 <div class="carousel-item {% if forloop.counter == 1 %}active{% endif %}">
-                    <img class="d-block w-100" style="width: 2%; height: 28rem" src="{% static image %}"
+                    <img class="d-block w-100" style="width: 2%; height: 28rem" src="{{ carousel.image_carousel.url }}"
                          alt="">
                 </div>
                 {% endfor %}
